+++ task/repository.go
@@ -1,7 +1,7 @@
 package task
 
 type Repository interface {
-	Create(query string) (int, error)
+	Create(query string) (int64, error)
 	Read(query string) (Scanner, error)
 	Update(query string) error
 	Delete(query string) error
